Add main logic app class with code initializing global drinks list
